# docs/spec/openapi.yml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: Percona Everest schema
  description: |
    # Authentication
    All requests to Everest API require `Authorization: Bearer <token>` header with a valid token in plain-text.
    
    The token can be obtained by using `everestctl token reset` which resets the token and prints it to the screen.
...
paths:
...
  '/namespaces/{namespace}/database-clusters':
    post:
      tags:
        - databaseCluster
      summary: Create a database cluster
      description: Create a database cluster
      operationId: createDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCluster'
    get:
      tags:
        - databaseCluster
      summary: List of the created database clusters
      description: List of the created database clusters
      operationId: listDatabaseClusters
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/namespaces/{namespace}/database-clusters/{name}':
    get:
      tags:
        - databaseCluster
      summary: Get the specified database cluster
      description: Get the specified database cluster
      operationId: getDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - databaseCluster
      summary: Replace the specified database cluster
      description: Replace the specified database cluster
      operationId: updateDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCluster'
    delete:
      tags:
        - databaseCluster
      summary: Delete the specified database cluster
      description: Delete the specified database cluster
      operationId: deleteDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/namespaces/{namespace}/database-clusters/{name}/backups':
    get:
      tags:
        - databaseClusterBackup
      summary: List of the created database cluster backups
      description: List of the created database cluster backups
      operationId: listDatabaseClusterBackups
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterBackupList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/namespaces/{namespace}/database-clusters/{name}/credentials':
    get:
      tags:
        - databaseCluster
      summary: Get the specified database cluster credentials
      description: Get the specified database cluster credentials
      operationId: getDatabaseClusterCredentials
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterCredential'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/namespaces/{namespace}/database-clusters/{name}/pitr':
    get:
      tags:
        - databaseCluster
      summary: Get the Point-in-Time related data for the specified database cluster
      description: Get the Point-in-Time related data for the specified database cluster
      operationId: getDatabaseClusterPitr
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterPitr'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
...
    DatabaseClusterCredential:
      type: object
      description: kubernetes object
      properties:
        username:
          type: string
          example: root
        password:
          type: string
          example: root
...
    DatabaseClusterList:
      description: DatabaseClusterList is an object that contains the list of the existing database clusters.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
                  of an object. Servers should convert recognized schemas to the latest
                  internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this
                  object represents. Servers may infer this from the endpoint the client
                  submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/DatabaseCluster'
        metadata:
          type: object
      type: object
...
    DatabaseCluster:
      description: DatabaseCluster is the Schema for the databaseclusters API.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        metadata:
          type: object
        spec:
          description: DatabaseClusterSpec defines the desired state of DatabaseCluster.
          properties:
            allowUnsafeConfiguration:
              description: AllowUnsafeConfiguration field used to ensure that the user can create configurations unfit for production use.
              type: boolean
            backup:
              description: Backup is the backup specification
              properties:
                enabled:
                  description: Enabled is a flag to enable backups
                  type: boolean
                pitr:
                  description: PITR is the configuration of the point in time recovery
                  properties:
                    backupStorageName:
                      description: BackupStorageName is the name of the BackupStorage where the PITR is enabled
                      type: string
                    enabled:
                      description: Enabled is a flag to enable PITR
                      type: boolean
                    uploadIntervalSec:
                      description: UploadIntervalSec number of seconds between the binlogs uploads
                      type: integer
                  required:
                    - enabled
                  type: object
                schedules:
                  description: Schedules is a list of backup schedules
                  items:
                    description: BackupSchedule is the backup schedule configuration.
                    properties:
                      backupStorageName:
                        description: BackupStorageName is the name of the BackupStorage CR that defines the storage location
                        type: string
                      enabled:
                        description: Enabled is a flag to enable the schedule
                        type: boolean
                      name:
                        description: Name is the name of the schedule
                        type: string
                      retentionCopies:
                        description: RetentionCopies is the number of backup copies to retain
                        format: int32
                        type: integer
                      schedule:
                        description: Schedule is the cron schedule
                        type: string
                    required:
                      - backupStorageName
                      - enabled
                      - name
                      - schedule
                    type: object
                  type: array
              required:
                - enabled
              type: object
            dataSource:
              description: DataSource defines a data source for bootstraping a new cluster
              properties:
                backupSource:
                  description: BackupSource is the backup source to restore from
                  properties:
                    backupStorageName:
                      description: BackupStorageName is the name of the BackupStorage used for backups.
                      type: string
                    path:
                      description: Path is the path to the backup file/directory.
                      type: string
                  required:
                    - backupStorageName
                    - path
                  type: object
                dbClusterBackupName:
                  description: DBClusterBackupName is the name of the DB cluster backup to restore from
                  type: string
                pitr:
                  description: PITR is the point-in-time recovery configuration
                  properties:
                    date:
                      description: 'Date is the UTC date to recover to. The accepted format: "2006-01-02T15:04:05Z".'
                      type: string
                    type:
                      default: date
                      description: Type is the type of recovery.
                      enum:
                        - date
                        - latest
                      type: string
                  type: object
              type: object
            engine:
              description: Engine is the database engine specification
              properties:
                config:
                  description: Config is the engine configuration
                  type: string
                replicas:
                  description: Replicas is the number of engine replicas
                  format: int32
                  minimum: 1
                  type: integer
                resources:
                  description: Resources are the resource limits for each engine replica. If not set, resource limits are not imposed
                  properties:
                    cpu:
                      anyOf:
                        - type: integer
                        - type: string
                      description: CPU is the CPU resource requirements
                      pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                      x-kubernetes-int-or-string: true
                    memory:
                      anyOf:
                        - type: integer
                        - type: string
                      description: Memory is the memory resource requirements
                      pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                      x-kubernetes-int-or-string: true
                  type: object
                storage:
                  description: Storage is the engine storage configuration
                  properties:
                    class:
                      description: Class is the storage class to use for the persistent volume claim
                      type: string
                    size:
                      anyOf:
                        - type: integer
                        - type: string
                      description: Size is the size of the persistent volume claim
                      pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                      x-kubernetes-int-or-string: true
                  required:
                    - size
                  type: object
                type:
                  description: Type is the engine type
                  enum:
                    - pxc
                    - postgresql
                    - psmdb
                  type: string
                userSecretsName:
                  description: UserSecretsName is the name of the secret containing the user secrets
                  type: string
                version:
                  description: Version is the engine version
                  type: string
              required:
                - storage
                - type
              type: object
            monitoring:
              description: Monitoring is the monitoring configuration
              properties:
                monitoringConfigName:
                  description: MonitoringConfigName is the name of a monitoringConfig CR.
                  type: string
                resources:
                  description: Resources defines resource limitations for the monitoring.
                  properties:
                    claims:
                      description: "Claims lists the names of resources, defined in spec.resourceClaims, that are used by this container. \n This is an alpha field and requires enabling the DynamicResourceAllocation feature gate. \n This field is immutable. It can only be set for containers."
                      items:
                        description: ResourceClaim references one entry in PodSpec.ResourceClaims.
                        properties:
                          name:
                            description: Name must match the name of one entry in pod.spec.resourceClaims of the Pod where this field is used. It makes that resource available inside a container.
                            type: string
                        required:
                          - name
                        type: object
                      type: array
                      x-kubernetes-list-map-keys:
                        - name
                      x-kubernetes-list-type: map
                    limits:
                      additionalProperties:
                        anyOf:
                          - type: integer
                          - type: string
                        pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                        x-kubernetes-int-or-string: true
                      description: 'Limits describes the maximum amount of compute resources allowed. More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/'
                      type: object
                    requests:
                      additionalProperties:
                        anyOf:
                          - type: integer
                          - type: string
                        pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                        x-kubernetes-int-or-string: true
                      description: 'Requests describes the minimum amount of compute resources required. If Requests is omitted for a container, it defaults to Limits if that is explicitly specified, otherwise to an implementation-defined value. Requests cannot exceed Limits. More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/'
                      type: object
                  type: object
              type: object
            paused:
              description: Paused is a flag to stop the cluster
              type: boolean
            proxy:
              description: Proxy is the proxy specification. If not set, an appropriate proxy specification will be applied for the given engine. A common use case for setting this field is to control the external access to the database cluster.
              properties:
                config:
                  description: Config is the proxy configuration
                  type: string
                expose:
                  description: Expose is the proxy expose configuration
                  properties:
                    ipSourceRanges:
                      description: IPSourceRanges is the list of IP source ranges (CIDR notation) to allow access from. If not set, there is no limitations
                      items:
                        description: IPSourceRange represents IP addresses in CIDR notation or without a netmask.
                        type: string
                      type: array
                    type:
                      default: internal
                      description: Type is the expose type, can be internal or external
                      enum:
                        - internal
                        - external
                      type: string
                  type: object
                replicas:
                  description: Replicas is the number of proxy replicas
                  format: int32
                  minimum: 1
                  type: integer
                resources:
                  description: Resources are the resource limits for each proxy replica. If not set, resource limits are not imposed
                  properties:
                    cpu:
                      anyOf:
                        - type: integer
                        - type: string
                      description: CPU is the CPU resource requirements
                      pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                      x-kubernetes-int-or-string: true
                    memory:
                      anyOf:
                        - type: integer
                        - type: string
                      description: Memory is the memory resource requirements
                      pattern: ^(\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))(([KMGTPE]i)|[numkMGTPE]|([eE](\+|-)?(([0-9]+(\.[0-9]*)?)|(\.[0-9]+))))?$
                      x-kubernetes-int-or-string: true
                  type: object
                type:
                  description: Type is the proxy type
                  enum:
                    - mongos
                    - haproxy
                    - proxysql
                    - pgbouncer
                  type: string
              type: object
          required:
            - engine
          type: object
        status:
          description: DatabaseClusterStatus defines the observed state of DatabaseCluster.
          properties:
            activeStorage:
              description: ActiveStorage is the storage used in cluster (psmdb only)
              type: string
            hostname:
              description: Hostname is the hostname where the cluster can be reached
              type: string
            message:
              description: Message is extra information about the cluster
              type: string
            port:
              description: Port is the port where the cluster can be reached
              format: int32
              type: integer
            ready:
              description: Ready is the number of ready pods
              format: int32
              type: integer
            size:
              description: Size is the total number of pods
              format: int32
              type: integer
            status:
              description: Status is the status of the cluster
              type: string
          type: object
      type: object
    # -------------------------
    # DatabaseClusterBackup spec is imported from the everest operator,
    # please do not edit the definition below manually
    # -------------------------
...